+++ src/Components/WebView/Headless/src/Document/ComponentNode.cs
@@ -1,0 +1,12 @@
+﻿namespace Microsoft.AspNetCore.Components.WebView.Headless.Document
+{
+    internal class ComponentNode : ContainerNode
+    {
+        public ComponentNode(int componentId)
+        {
+            ComponentId = componentId;
+        }
+
+        public int ComponentId { get; }
+    }
+}
