+++ src/Components/WebView/Headless/src/Document/HeadlessNode.cs
@@ -1,0 +1,7 @@
+﻿namespace Microsoft.AspNetCore.Components.WebView.Headless.Document
+{
+    internal class HeadlessNode
+    {
+        public ContainerNode Parent { get; set; }
+    }
+}
