+++ src/Components/WebView/Headless/src/Document/LogicalContainerNode.cs
@@ -1,0 +1,6 @@
+﻿namespace Microsoft.AspNetCore.Components.WebView.Headless.Document
+{
+    internal class LogicalContainerNode : ContainerNode
+    {
+    }
+}
