+++ src/Components/WebView/Headless/src/Document/MarkupNode.cs
@@ -1,0 +1,12 @@
+﻿namespace Microsoft.AspNetCore.Components.WebView.Headless.Document
+{
+    internal class MarkupNode : HeadlessNode
+    {
+        public MarkupNode(string markupContent)
+        {
+            Content = markupContent;
+        }
+
+        public string Content { get; }
+    }
+}
