+++ src/Components/WebView/Headless/src/HeadlessWebViewServiceCollectionExtensions.cs
@@ -1,4 +1,4 @@
-﻿using Microsoft.AspNetCore.Components.Routing;
+using Microsoft.AspNetCore.Components.Routing;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.JSInterop;
 
@@ -9,10 +9,7 @@ namespace Microsoft.AspNetCore.Components.WebView.Headless
         public static IServiceCollection AddHeadlessWebView(this IServiceCollection services)
         {
             services.AddBlazorWebView();
-            services.AddScoped<WebViewHost>();
-            services.AddScoped<IJSRuntime, WebViewJSRuntime>();
-            services.AddScoped<INavigationInterception, WebViewNavigationInterception>();
-            services.AddScoped<NavigationManager, WebViewNavigationManager>();
+            services.AddScoped<WebViewHost, HeadlessWebViewHost>();
 
             return services;
         }
