+++ src/Components/WebView/Headless/src/Microsoft.AspNetCore.Components.WebView.Headless.csproj
@@ -20,9 +20,4 @@
     <Reference Include="Microsoft.Extensions.Logging" />
   </ItemGroup>
 
-  <ItemGroup>
-    <Folder Include="Document\" />
-    <Compile Include="$(ComponentsSharedSourceRoot)src\WebEventData.cs" />
-  </ItemGroup>
-
 </Project>
