+++ src/Components/WebView/Hosting/src/BlazorWebViewServiceCollectionExtensions.cs
@@ -1,9 +1,6 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+using Microsoft.AspNetCore.Components.Routing;
 using Microsoft.Extensions.DependencyInjection;
+using Microsoft.JSInterop;
 
 namespace Microsoft.AspNetCore.Components.WebView
 {
@@ -12,6 +9,10 @@ namespace Microsoft.AspNetCore.Components.WebView
         public static IServiceCollection AddBlazorWebView(this IServiceCollection services)
         {
             services.AddLogging();
+            services.AddScoped<IJSRuntime, WebViewJSRuntime>();
+            services.AddScoped<INavigationInterception, WebViewNavigationInterception>();
+            services.AddScoped<NavigationManager, WebViewNavigationManager>();
+
             return services;
         }
     }
